requirements.txt
appdirs>=1.4.4
audioread>=3.0.0
cached-property>=1.5.2
certifi>=2023.5.7
cffi>=1.15.1
charset-normalizer>=2.0.12
cycler>=0.11.0
Cython>=3.0.0
decorator>=5.1.1
h5py>=3.1.0
idna>=3.4
joblib>=1.1.1
kiwisolver>=1.3.1
librosa>=0.9.2
llvmlite>=0.36.0
matplotlib>=3.3.4
numba>=0.53.1
numpy>=1.19.5
packaging>=21.3
pandas>=1.1.5
Pillow>=8.4.0
pooch>=1.6.0
pybind11>=2.11.1
pycparser>=2.21
pyparsing>=3.1.0
pyroomacoustics>=0.7.3
python-dateutil>=2.8.2
pytz>=2024.1
requests>=2.27.1
resampy>=0.4.2
scikit-learn>=0.24.2
scipy>=1.5.4
six>=1.16.0
soundfile>=0.12.1
threadpoolctl>=3.1.0
torch>=1.10.1
torchaudio>=0.10.1
typing_extensions>=4.1.1
urllib3>=1.26.16
Pyyaml
